fix+perf(core): remove dead cast, zero-alloc LIKE hot-path, robust streaming test, O(n log n) index backfill (#805)
* fix(core): remove redundant cast, zero-alloc LIKE hot-path, robust streaming test

- distance.rs: drop `as f32` on hamming_distance_native (already returns f32)
- filter/matching.rs: avoid String allocation on the LIKE case-sensitive path;
  only allocate for to_lowercase() when case_insensitive=true — ILIKE is the
  minority case; LIKE now zero-alloc relative to the input lifetime
- search_tests.rs: replace get()-based poll with search()-based poll in
  test_enable_streaming_sends_and_searches; get() confirms storage visibility
  but not HNSW index readiness, which was the source of the intermittent
  assertion failure at line 408 under heavy parallel test load

All 4759 unit tests pass.


* perf(core): O(n log n) secondary-index backfill via sort+dedup

Replace the O(n²) per-insertion contains() guard in backfill_single_payload
with a single bulk sort_unstable()+dedup() pass after the loop.

For a low-cardinality field with k distinct IDs per bucket the old code was
O(0+1+2+…+(k-1))=O(k²); the new code is O(n log n) over all buckets combined,
which is the lower bound for a comparison sort and typically 10-100× faster on
realistic collection sizes.


* style(core): fix rustfmt violations flagged by CI

Two lines exceeded rustfmt's width limit:
- filter/matching.rs: split like_match_impl() call across three lines
- search_tests.rs: split assert!() call across three lines


* test(streaming): preserve explicit timeout diagnostic in search-poll loop

The previous version of the polling loop broke out on either readiness
or deadline expiry, then asserted `!results.is_empty()` afterward. If
the deadline expired before HNSW indexed the streamed points, the
failure surfaced through the weaker "streamed points should be
searchable" message — the original "timed out waiting for stream flush"
diagnostic was lost.

Track timeout explicitly and assert on it first, so a CI flake clearly
distinguishes "indexing slow / deadline hit" from "wrong top result".
No behaviour change on the happy path.


---------

+++ crates/velesdb-core/src/distance.rs
@@ -102,7 +102,7 @@ impl DistanceMetric {
             Self::Cosine => simd_native::cosine_similarity_native(a, b),
             Self::Euclidean => simd_native::euclidean_native(a, b),
             Self::DotProduct => simd_native::dot_product_native(a, b),
-            Self::Hamming => simd_native::hamming_distance_native(a, b) as f32,
+            Self::Hamming => simd_native::hamming_distance_native(a, b),
             Self::Jaccard => simd_native::jaccard_similarity_native(a, b),
         }
     }
